only review fixes

+++ tasks/all/zinoviev_a_convex_hull_components/perf_tests/perf_all.cpp
@@ -38,7 +38,7 @@ TEST(zinoviev_a_convex_hull_components_all, test_pipeline_run) {
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
   const int size = 5000;
-  std::vector<int> input(size * size, 0);
+  std::vector<int> input(static_cast<size_t>(size) * static_cast<size_t>(size), 0);
 
   for (int i = 0; i < size; ++i) {
     input[i] = 1;
@@ -88,7 +88,7 @@ TEST(zinoviev_a_convex_hull_components_all, test_task_run) {
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
   const int size = 5000;
-  std::vector<int> input(size * size, 0);
+  std::vector<int> input(static_cast<size_t>(size) * static_cast<size_t>(size), 0);
 
   for (int i = 0; i < size; ++i) {
     input[i] = 1;
